following and unfollowing listing functionnality ok, following display in dasboard WIP

+++ followings/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth.decorators import login_required
 from .models import Following
 from django.contrib import messages
 from django.http import JsonResponse
+from django.urls import reverse
 # Create your views here.
 
 
@@ -18,11 +19,14 @@ def follow(request):
         following = Following(user_id=user_id, listing_id=listing_id, listing=listing)
         has_followed = Following.objects.all().filter(listing_id=listing_id, user_id=user_id)
         if has_followed:
-            return JsonResponse({"message": "You already follow this listing!", "result": "error", "class": "danger"}, status=200)
+            return JsonResponse({"message": "You already follow this listing!", "result": "error", "class": "danger"},
+                                status=200)
         else:
             following.save()
 
-        return JsonResponse({"message": "You now follow this listing!", "result": "success", "class": "success"}, status=200)
+        return JsonResponse({"message": "You now follow this listing!", "result": "success", "class": "success",
+                             "route": reverse('unfollow')},
+                            status=200)
 
 
 @login_required
@@ -30,9 +34,11 @@ def unfollow(request):
     """
     allows user to remove listing's following
     """
-    user_id = request.user.id
-    listing_id = request.POST['listing_id']
-    following = Following.objects.all().filter(listing_id=listing_id, user_id=user_id)
-    if following:
-        following.delete()
-    return redirect('/listings/' + listing_id)
+    if request.method == 'POST' and request.is_ajax:
+        user_id = request.user.id
+        listing_id = request.POST['listing_id']
+        following = Following.objects.all().filter(listing_id=listing_id, user_id=user_id)
+        if following:
+            following.delete()
+            return JsonResponse({"message": "You are no more following this listing!", "result": "success",
+                                 "class": "info", "route": reverse('follow')}, status=200)
